[6.1] delete location frontend, only deleting non default locationtags

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/StorageServiceImpl.java
@@ -246,6 +246,10 @@ public class StorageServiceImpl implements StorageService {
     @Override
     public void deleteLocation(Long id) {
         LOGGER.debug("Deleting a location");
+        LocationClass locationToDelete = locationRepository.getById(id);
+        if(locationToDelete.getStorageId() == null) {
+            throw new ServiceException("location cannot be delete!");
+        }
         locationRepository.deleteById(id);
     }
